# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Timetracker API",
    "description": "OpenaAPI SPEC oF my fancy timetracker app",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "6.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:30000/api"
    },
    {
      "url": "https://localhost:30443/api"
    },
    {
      "url": "https://nobio.myhome-server.de/api"
    },
    {
      "url": "https://timetracker-ui.web.app/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "keyAuth": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication & Authorization"
        ],
        "summary": "Login of a user",
        "description": "A user (username, password) will be authenticated; a jwt access token (5min expiry) and a refresh token will be returned.",
        "requestBody": {
          "description": "The User object to be authenticated",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "tester",
                    "maxLength": 128,
                    "pattern": "^[\\w-.]"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "pattern": "^[\\w-.]",
                    "example": "psiodjkÜsö93JKK=w8dj",
                    "maxLength": 20
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "access and refresh tokens are created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "the accees JWT to be used to access",
                      "example": "eyJhbJ9.eyJuYW1lIjoibm4OTg0Mn0.kgR2dN1VsZ_1Vd8\","
                    },
                    "refreshToken": {
                      "type": "string",
                      "description": "refresh JWT",
                      "example": "eyIjshJ9.eyJOOksW1lIjdasdihg0Mn0.kgR2dNsda2d8\","
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized; usually when no password was provided or when user is not registered or when password was incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "error message",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "User"
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication & Authorization"
        ],
        "summary": "logout and destroy tokens; does not need bearer token",
        "requestBody": {
          "description": "long living refresh token will be invalidated (short living access token might still live until expiry)",
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "token"
                ],
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "JWT",
                    "pattern": "^[A-Za-z0-9-_]*\\.[A-Za-z0-9-_]*\\.[A-Za-z0-9-_]*$",
                    "maxLength": 256,
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invalidates the refresh token of a user"
          }
        }
      }
    },
    "/auth/token": {
      "post": {
        "tags": [
          "Authentication & Authorization"
        ],
        "summary": "Refresh the access token using the refresh token",
        "requestBody": {
          "description": "long living refresh token will be invalidated (short living access token might still live until expiry)",
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "token"
                ],
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "JWT",
                    "pattern": "^[A-Za-z0-9-_]*\\.[A-Za-z0-9-_]*\\.[A-Za-z0-9-_]*$",
                    "maxLength": 256,
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "new access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "new accees JWT to be used to access",
                      "example": "eyJhbJ9.eyJuYW1lIjoibm4OTg0Mn0.kgR2dN1VsZ_1Vd8\","
                    },
                    "user": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "name": {
                          "type": "string",
                          "description": "user name linked to this token",
                          "example": "user@university.edu"
                        },
                        "iat": {
                          "type": "number",
                          "description": "generation timetamp",
                          "example": 1618390174
                        },
                        "exp": {
                          "type": "number",
                          "description": "expire timetamp",
                          "example": 1618390174
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "refresh token is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "error message",
                      "example": "Unauthorized (invalid refresh token)"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/entries": {
      "get": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Read all time entries",
        "description": "Reads all time entries;\nread entries by date getAllByDate -> List of Time Entry of the given day;\nRead busy time getBusyTime -> Busy Time of the given day (in ms)\n",
        "parameters": [
          {
            "name": "busy",
            "in": "query",
            "description": "Zeitstempel",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[\\w-.]"
            }
          },
          {
            "name": "dt",
            "in": "query",
            "description": "Zeitstempel",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[\\w-.]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All (or filtered) time entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "maxItems": 999999,
                  "items": {
                    "required": [
                      "direction",
                      "entry_date"
                    ],
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "entry_date": {
                        "type": "string",
                        "description": "The date this time entry represents",
                        "example": "2000-01-01T10:15:02.151Z",
                        "format": "date-time"
                      },
                      "last_changed": {
                        "type": "string",
                        "description": "Date of last change on this time entry",
                        "example": "2000-01-01T10:15:02.151Z",
                        "format": "date-time"
                      },
                      "direction": {
                        "type": "string",
                        "description": "direction of this time entry; representing, if the time entry means a coming or going event:\n * `go` - leaving\n * `enter` - entering\n",
                        "example": "enter",
                        "enum": [
                          "enter",
                          "go"
                        ]
                      },
                      "longitude": {
                        "type": "number",
                        "description": "the longitude geo coordinate",
                        "example": 10.875482
                      },
                      "latitude": {
                        "type": "number",
                        "description": "the latitude geo coordinate",
                        "example": 49.514135
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        }
      },
      "post": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Create new time entry",
        "description": "Create new time entry in system",
        "requestBody": {
          "description": "Time entry that we want to create",
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "direction",
                  "entry_date"
                ],
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "entry_date": {
                    "type": "string",
                    "description": "The date this time entry represents",
                    "example": "2000-01-01T10:15:02.151Z",
                    "format": "date-time"
                  },
                  "last_changed": {
                    "type": "string",
                    "description": "Date of last change on this time entry",
                    "example": "2000-01-01T10:15:02.151Z",
                    "format": "date-time"
                  },
                  "direction": {
                    "type": "string",
                    "description": "direction of this time entry; representing, if the time entry means a coming or going event:\n * `go` - leaving\n * `enter` - entering\n",
                    "example": "enter",
                    "enum": [
                      "enter",
                      "go"
                    ]
                  },
                  "longitude": {
                    "type": "number",
                    "description": "the longitude geo coordinate",
                    "example": 10.875482
                  },
                  "latitude": {
                    "type": "number",
                    "description": "the latitude geo coordinate",
                    "example": 49.514135
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "New entry is created",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "direction",
                    "entry_date"
                  ],
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "entry_date": {
                      "type": "string",
                      "description": "The date this time entry represents",
                      "example": "2000-01-01T10:15:02.151Z",
                      "format": "date-time"
                    },
                    "last_changed": {
                      "type": "string",
                      "description": "Date of last change on this time entry",
                      "example": "2000-01-01T10:15:02.151Z",
                      "format": "date-time"
                    },
                    "direction": {
                      "type": "string",
                      "description": "direction of this time entry; representing, if the time entry means a coming or going event:\n * `go` - leaving\n * `enter` - entering\n",
                      "example": "enter",
                      "enum": [
                        "enter",
                        "go"
                      ]
                    },
                    "longitude": {
                      "type": "number",
                      "description": "the longitude geo coordinate",
                      "example": 10.875482
                    },
                    "latitude": {
                      "type": "number",
                      "description": "the latitude geo coordinate",
                      "example": 49.514135
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        },
        "x-codegen-request-body-name": "timeentry"
      }
    },
...
    "/entries/dump": {
      "post": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Backup data to file system",
        "description": "dump the mongodb to the local file system in order to be restored if needed",
        "responses": {
          "200": {
            "description": "data dumped to file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        }
      }
    },
    "/entries/restore": {
      "post": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Restore data",
        "description": "Restore data from file which has been dumped by dump method",
        "responses": {
          "200": {
            "description": "data sucessfully restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        }
      }
    },
    "/entries/backup": {
      "post": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Backup data to database",
        "description": "backup data in an extra backup table",
        "responses": {
          "200": {
            "description": "data saved to database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          },
          "500": {
            "description": "something went wrong...",
            "content": {}
          }
        }
      }
    },
    "/entries/error/evaluate": {
      "post": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Evaluate time entries",
        "description": "Evaluate time entries on their consistency (completeness, order, etc.). As a result, the failure data will be put to an own document in mongodb. This data could be inquired by using the endpoint /api/entries/error/dates",
        "responses": {
          "200": {
            "description": "data saved to database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          },
          "500": {
            "description": "something went wrong...",
            "content": {}
          }
        }
      }
    },
    "/entries/error/dates": {
      "get": {
        "tags": [
          "TimeEntries"
        ],
        "summary": "Inquire the error time entries",
        "description": "Read all dates with error type that was written to database by endpoint /api/entries/error/evaluate",
        "responses": {
          "200": {
            "description": "data saved to database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "maxItems": 999999,
                  "items": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "error-date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2000-01-01T10:15:02.151Z"
                      },
                      "error-type": {
                        "type": "string",
                        "maxLength": 30,
                        "description": "Type of the error - INCOMPLETE means that entries are missing and WRONG_ORDER means that the order of go and enter is not correctly booked",
                        "example": "INCOMPLETE or WRONG_ORDER"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          },
          "500": {
            "description": "something went wrong...",
            "content": {}
          }
        }
      }
    },
...
    "/geotrack/metadata": {
      "get": {
        "tags": [
          "Geo Tracking"
        ],
        "summary": "meta data about the track for the given time span",
        "description": "Returns som meta data about the track for the given time span",
        "parameters": [
          {
            "name": "dateStart",
            "in": "query",
            "description": "start date of the time range; format YYYY-MM-DD like 2020-03-09",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateEnd",
            "in": "query",
            "description": "end date of the time range; format YYYY-MM-DD like 2020-03-10",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "delivered a valid result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "size": {
                      "type": "number",
                      "description": "number of geo track entries for the given time span",
                      "example": 210
                    },
                    "totalDistance": {
                      "type": "number",
                      "description": "distance in meter of the given time span",
                      "example": 21232
                    },
                    "accuracy": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "mean": {
                          "type": "number",
                          "description": "mean of accuracies in meters of all geo locations",
                          "example": 114.232
                        },
                        "variance": {
                          "type": "number",
                          "description": "variance of accuracies in meters of all geo locations",
                          "example": 11.031
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        }
      }
    },
...
    "/toggles/name/{name}": {
      "get": {
        "tags": [
          "Toggles"
        ],
        "summary": "read one toggle by its name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of a toggle that we want to find",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[\\w-.]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All toggles status",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "name",
                    "notification",
                    "toggle"
                  ],
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "__TEST__",
                      "maxLength": 128,
                      "pattern": "^[\\w-.]"
                    },
                    "toggle": {
                      "type": "string",
                      "example": "false",
                      "maxLength": 128,
                      "pattern": "^[\\w-.]"
                    },
                    "notification": {
                      "type": "string",
                      "example": "some text",
                      "maxLength": 128,
                      "pattern": "^[\\w-.]"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access token is missing or invalid; try to refresh your token"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "keyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api-key",
        "description": "All requests must include the `api-key` header containing your account ID."
      }
    },
...
  }
}